Mise à jour de la fonction extrait ajout de fonctionnalité URL content

+++ crawl/admin/extract.php
@@ -1,5 +1,7 @@
 <?php
 
+
+   
     error_reporting (E_ALL ^ E_NOTICE ^ E_WARNING); 
     include "configSearch/cookie.php";
     include "auth.php";
@@ -22,15 +24,19 @@
     }
 
     mysql_query("ALTER TABLE ".$mysql_table_prefix."links ADD $column MEDIUMTEXT");
-
-    $sql = "SELECT link_id,fullhtml FROM ".$mysql_table_prefix."links WHERE site_id='$site'";
+//WHERE site_id='$site'
+    $sql = "SELECT link_id,fullhtml FROM ".$mysql_table_prefix."links ";
     if ($in!="") $sql .= " AND url LIKE '%$in%'";
-    if ($out!="") $sql .= " AND url NOT LIKE '%$out%'";  
+    if ($out!="") $sql .= " AND url NOT LIKE '%$out%'";
+    //$sql .= " AND link_id > 31000";  
     $result = mysql_query($sql);
 
+echo '::'.$actual_column.'::';
     while ($tab = mysql_fetch_array($result)){
 
-        $html = $tab['fullhtml'];
+        if($actual_column){$html = $tab[$actual_column];}else{$html = $tab['fullhtml'];}
+        
+        
         if ($tag_name!=""){           
             // Cas intérieur <tag>
             $doc = new DOMDocument();
@@ -66,32 +72,45 @@
         }
 
         if ($html!="" && $html!=chr(10)){
+        
+          if($keep_html=='on'){
 
-            $partialtxt = $html;        
-
-            $partialtxt = preg_replace("/<link rel[^<>]*>/i", " ", $partialtxt);
-            $partialtxt = preg_replace("@<!--sphider_noindex-->.*?<!--\/sphider_noindex-->@si", " ",$partialtxt);    
-            $partialtxt = preg_replace("@<!--.*?-->@si", " ",$partialtxt);    
-            $partialtxt = preg_replace("@<script[^>]*?>.*?</script>@si", " ",$partialtxt);
-
-            $regs = Array ();
-            if (preg_match("@<title *>(.*?)<\/title*>@si", $partialtxt, $regs)) {
-                $partialtxt = str_replace($regs[0], "", $partialtxt);
-            }
-
-            $partialtxt = preg_replace("@<style[^>]*>.*?<\/style>@si", " ", $partialtxt);               
-
-            // HTML tags
-            $partialtxt = preg_replace("/&lt;(\/?[^(&gt;)]+)&gt;/", "<\\1>", $partialtxt);   
-            //create spaces between tags, so that removing tags doesnt concatenate strings
-            $partialtxt = preg_replace("/\<(\/?[^\>]+)\>/", "\\0 ", $partialtxt);       
-            $partialtxt = strip_tags($partialtxt);                  
-            $partialtxt = htmlToISO($partialtxt);
-
-            mysql_query("UPDATE ".$mysql_table_prefix."links SET $column='$partialtxt' WHERE link_id='$tab[link_id]'");
-            echo "<b>".mysql_error()."</b>"; 
-            echo "<br><br><b>$tab[link_id]</b>",$partialtxt;
-
+			            $partialtxt = $html;        
+			            $partialtxt = htmlToISO($partialtxt);
+			
+			            mysql_query("UPDATE ".$mysql_table_prefix."links SET $column='$partialtxt' WHERE link_id='$tab[link_id]'");
+			            echo "<b>".mysql_error()."</b>"; 
+			            echo "<br><br><b>HTML : $tab[link_id]</b> ",$partialtxt;
+			               
+		   }else{
+		   
+		   			   $partialtxt = $html;        
+		   
+		               $partialtxt = preg_replace("/<link rel[^<>]*>/i", " ", $partialtxt);
+		               $partialtxt = preg_replace("@<!--sphider_noindex-->.*?<!--\/sphider_noindex-->@si", " ",$partialtxt);    
+		               $partialtxt = preg_replace("@<!--.*?-->@si", " ",$partialtxt);    
+		               $partialtxt = preg_replace("@<script[^>]*?>.*?</script>@si", " ",$partialtxt);
+		   
+		               $regs = Array ();
+		               if (preg_match("@<title *>(.*?)<\/title*>@si", $partialtxt, $regs)) {
+		                   $partialtxt = str_replace($regs[0], "", $partialtxt);
+		               }
+		   
+		               $partialtxt = preg_replace("@<style[^>]*>.*?<\/style>@si", " ", $partialtxt);               
+		   
+		               // HTML tags
+		               $partialtxt = preg_replace("/&lt;(\/?[^(&gt;)]+)&gt;/", "<\\1>", $partialtxt);   
+		               //create spaces between tags, so that removing tags doesnt concatenate strings
+		               $partialtxt = preg_replace("/\<(\/?[^\>]+)\>/", "\\0 ", $partialtxt);       
+		               $partialtxt = strip_tags($partialtxt);                  
+		               $partialtxt = htmlToISO($partialtxt);
+		   
+		               mysql_query("UPDATE ".$mysql_table_prefix."links SET $column='$partialtxt' WHERE link_id='$tab[link_id]'");
+		               echo "<b>".mysql_error()."</b>"; 
+		               echo "<br><br><b>$tab[link_id]</b>",$partialtxt;
+		   		   
+		   }
+		
         }
 
     }
